-information table updated -zoom + pan implemented (new offset after zoom not implemented yet)
git-svn-id: file:///srv/local/data/cleanRepo/trunk@17332 a0517364-e873-4167-b1c8-416daa5a6b91

+++ playgrounds/wdoering/src/main/java/playground/wdoering/debugvisualization/gui/GUI.java
@@ -47,7 +47,7 @@ public class GUI extends JFrame {
 		
 		//get toolbar and renderer (processing applet)
 		guiToolbar = new GUIToolbar(controller);
-		renderer = new P3DRenderer(controller, traceTimeRange, width, height, visualizationMode);
+		renderer = new P3DRenderer(controller, traceTimeRange, width, (int)(height), visualizationMode);
 		guiInfo = new GUIInfo(controller);
 		
 		//guiInfo.disableUpdate(true);
